consolidate TextViewUtils logic into TextViewAdapters
